Split out DSL for defining a grammar in code to a separate file
